Initial commit. Completed 'by-hand' data set.
This includes lots of tinkering to create a data set reflecting the simulations described by Lin et al. (1997, pp 424--6). It eventually became clear that the best way to start out would be to save every intermediate calculation as a variable so that it could be easily checked for errors. The finished product is the data set `dat` defined in lines 99--137, which is subject-interval level.